python-src/requirements.txt
# requirements.txt
fastapi
uvicorn
websockets
numpy
soundfile
torch
torchvision
transformers @ git+https://ghpxy.hwinzniej.top/https://github.com/huggingface/transformers@f742a644ca32e65758c3adb36225aef1731bd2a8
accelerate
qwen-omni-utils[decord]
webrtcvad
python-dotenv
pydantic-settings
scipy